initial commit in validation & viewmodel restructuring
