+++ src/main/java/com/webapp/TextBook/repository/data_access/User.java
@@ -94,8 +94,9 @@ public class User extends Person{
 
 
         // given: order of values goes pkey (username, password), userRole
-        this.password = (String)values[1];
-        this.userRole = UserRole.createUserRoleByNominalValue((String)values[2]).orElseThrow();
+        // ["blah", "blah", "username", "password", "user-role"]
+        this.password = (String)values[values.length - 2];
+        this.userRole = UserRole.createUserRoleByNominalValue((String)values[values.length - 1]).orElseThrow();
 
     }
 }
