+++ src/main/java/com/webapp/TextBook/validation/ApiValidation/dummyApiValidation.java
@@ -1,0 +1,4 @@
+package com.webapp.TextBook.validation.ApiValidation;
+
+public class dummyApiValidation {
+}
