+++ src/main/java/com/webapp/TextBook/validation/FormVallidation/loginUserInfoValidation/LogInUserInfoValidationInterface.java
@@ -1,4 +1,4 @@
-package com.webapp.TextBook.validation.loginUserInfoValidation;
+package com.webapp.TextBook.validation.FormVallidation.loginUserInfoValidation;
 import javax.validation.Constraint;
 import javax.validation.Payload;
 import java.lang.annotation.Documented;
@@ -8,21 +8,11 @@ import java.lang.annotation.Target;
 import static java.lang.annotation.ElementType.*;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
-
-
-import javax.validation.Constraint;
-import javax.validation.Payload;
-import java.lang.annotation.Documented;
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import static java.lang.annotation.ElementType.*;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
 @Target({ TYPE }) // add TYPE to be applicable to class
 @Retention(RUNTIME)
 @Constraint(validatedBy = LoginUserInfoValidatorImpl.class)
 @Documented
-public @interface  LogInUserInfoValidationInterface {
+public @interface LogInUserInfoValidationInterface {
     String message() default "{org.hibernate.validator.referenceguide.chapter06.CheckCase." +
             "message}";
