+++ src/main/java/com/webapp/TextBook/viewModel/apiViewModel/dummyApiViewModel.java
@@ -1,0 +1,4 @@
+package com.webapp.TextBook.viewModel.apiViewModel;
+
+public class dummyApiViewModel {
+}
