+++ src/main/java/com/webapp/TextBook/viewModel/formViewModel/dummyFormViewModel.java
@@ -1,0 +1,4 @@
+package com.webapp.TextBook.viewModel.formViewModel;
+
+public class dummyFormViewModel {
+}
