+++ src/main/java/com/webapp/TextBook/viewModel/shared/ApiVieweModel.java
@@ -1,0 +1,10 @@
+package com.webapp.TextBook.viewModel.shared;
+
+/**
+ * <h1>ApiViewModel</h1>
+ *
+ * <p>Place holder for shared-polymorphic grouping for all API View Models -- Form View Models excluded</p>
+ *
+ */
+public interface ApiVieweModel {
+}
