+++ src/main/java/com/webapp/TextBook/viewModel/shared/FormViewModel.java
@@ -1,0 +1,4 @@
+package com.webapp.TextBook.viewModel.shared;
+
+public interface FormViewModel {
+}
